Adds option to start server on startup
Implements a new setting that allows the chat server to start automatically when the application launches.

This improves user experience by removing the need to manually start the server each time the application is opened.
The setting is exposed in the settings view, allowing users to easily configure the desired behavior.